refactor: Update Pokemon component UI
The code changes in `pokemon.component.html` and `pokemon.component.ts` update the UI of the Pokemon component. The changes include adding a text-center class to the card div, styling the card text, adding a new div to display the number of curaciones, and updating the button styles. These UI improvements enhance the visual presentation and user experience of the Pokemon component.